4458FinalBackend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
psycopg2
fastapi
uvicorn[standard]
pydantic
azure-identity
azure-storage-queue
python-dotenv==0.19.1